[Decomp][Spell] Cohesion split: extract HandleAddModifier to SpellAuraAddModifier.cpp
Move the last remaining table-dispatched Aura::HandleXxx handler
(HandleAddModifier, for SPELL_AURA_ADD_FLAT_MODIFIER / ADD_PCT_MODIFIER) out of
SpellAuras.cpp into a new SpellAuraAddModifier.cpp translation unit.

This completes the Aura::HandleXxx handler decomposition: every aura handler
body now lives in a themed SpellAuraXxx.cpp TU and SpellAuras.cpp retains the
Aura/SpellAuraHolder class core plus the AuraHandler[] dispatch table, which
resolves all moved member methods by address at link time. Byte-identical move,
no behaviour change. SpellAuras.cpp 9040 -> 8986.
